backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
imagehash==4.3.1
Pillow==10.2.0
numpy==1.26.3
